% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapa_chr.R
\name{mapa_chr}
\alias{mapa_chr}
\title{mapa_chr}
\usage{
mapa_chr(cr = c(30.4, 19.7, 23.5, 18.6, 27), col_cr = "red",
  unidad = "bp")
}
\arguments{
\item{cr}{vector tiene que recibir un vector con valores numericos.}

\item{unidad}{unidad de los grupos (CM o Bp)}
}
\description{
crea un plot con grupos de anclaje
}
